itest: give the imported account keys neutral ownership
The deterministic imported-account XPub fixture lived in the account
manager's test file, but the Signer derivation cases need the same key
material to exercise imported-XPub public derivation. A component's test
file must not own a value another component depends on, so move the
fixture and its type to the neutral fixtures file.
